tambah mvc untuk profile dan kontak

+++ routes/web.php
@@ -1,9 +1,11 @@
 <?php
 
-use App\Http\Controllers\DashboardController;
-use App\Http\Controllers\HomeController;
-use App\Http\Controllers\ProdukController;
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\HomeController;
+use App\Http\Controllers\KontakController;
+use App\Http\Controllers\ProdukController;
+use App\Http\Controllers\ProfileController;
+use App\Http\Controllers\DashboardController;
 
 
 Route::get('/', [HomeController::class, 'index'])->name('home');
@@ -13,13 +15,15 @@ Route::get('/register', function () {
     return redirect('/login');
 });
 
-Route::get('/profile', function () {
-    return view('profile');
-})->name('profile');
+Route::get('/profile', [ProfileController::class, 'show'])->name('profile');
+// Route::get('/profile', function () {
+//     return view('profile');
+// })->name('profile');
 
-Route::get('/kontak', function () {
-    return view('kontak');
-})->name('kontak');
+Route::get('/kontak', [KontakController::class, 'show'])->name('kontak');
+// Route::get('/kontak', function () {
+//     return view('kontak');
+// })->name('kontak');
 
 
 Route::middleware('auth')->group(function () {
@@ -31,11 +35,19 @@ Route::middleware('auth')->group(function () {
     })->name('dashboard');
 
     // Halaman produk (CRUD)
-    Route::get('/home', [BerandaController::class, 'index']);
-    Route::get('/produk', [ProdukController::class, 'index'])->name('produk.index');
-    Route::post('/produk', [ProdukController::class, 'store'])->name('produk.store');
-    Route::put('/produk/{produk}', [ProdukController::class, 'update'])->name('produk.update');
-    Route::delete('/produk/{produk}', [ProdukController::class, 'destroy'])->name('produk.destroy');
+    // Route::get('/home', [BerandaController::class, 'index']);
+    Route::get('/admin/produk', [ProdukController::class, 'index'])->name('admin.produk.index');
+    Route::post('/admin/produk', [ProdukController::class, 'store'])->name('admin.produk.store');
+    Route::put('/admin/produk/{produk}', [ProdukController::class, 'update'])->name('admin.produk.update');
+    Route::delete('/admin/produk/{produk}', [ProdukController::class, 'destroy'])->name('admin.produk.destroy');
+
+    Route::get('/admin/kontak', [KontakController::class, 'form'])->name('admin.kontak.index');
+    Route::post('/admin/kontak', [KontakController::class, 'storeOrUpdate'])->name('admin.kontak.save');
+
+    // Admin profil
+    Route::get('/admin/profile', [ProfileController::class, 'form'])->name('admin.profile.index');
+    Route::post('/admin/profile', [ProfileController::class, 'storeOrUpdate'])->name('admin.profile.update');
+
 });
 
 require __DIR__.'/auth.php';
